API: deprecate undocumented functions
