+++ numpy/core/numerictypes.py
@@ -86,7 +86,7 @@ from .multiarray import (
         ndarray, array, dtype, datetime_data, datetime_as_string,
         busday_offset, busday_count, is_busday, busdaycalendar
         )
-from .._utils import set_module
+from .._utils import deprecate, set_module
 
 # we add more at the bottom
 __all__ = ['sctypeDict', 'sctypes',
@@ -129,6 +129,7 @@ genericTypeRank = ['bool', 'int8', 'uint8', 'int16', 'uint16',
                    'complex32', 'complex64', 'complex128', 'complex160',
                    'complex192', 'complex256', 'complex512', 'object']
 
+@deprecate
 @set_module('numpy')
 def maximum_sctype(t):
     """
@@ -245,7 +246,7 @@ def obj2sctype(rep, default=None):
 
     See Also
     --------
-    sctype2char, issctype, issubsctype, issubdtype, maximum_sctype
+    sctype2char, issctype, issubsctype, issubdtype
 
     Examples
     --------
