+++ numpy/core/tests/test_deprecations.py
@@ -777,7 +777,21 @@ class TestMathAlias(_DeprecationTestCase):
 class TestLibImports(_DeprecationTestCase):
     # Deprecated in Numpy 1.26.0, 2023-09
     def test_lib_functions_deprecation_call(self):
-        from numpy.lib import byte_bounds, safe_eval, who
+        from numpy import (
+            byte_bounds, safe_eval, who, recfromcsv, recfromtxt,
+            disp, get_array_wrap, maximum_sctype
+        )
+        from numpy.lib.tests.test_io import TextIO
+        
         self.assert_deprecated(lambda: byte_bounds(np.array([1])))
         self.assert_deprecated(lambda: safe_eval("None"))
         self.assert_deprecated(lambda: who())
+
+        data_gen = lambda: TextIO('A,B\n0,1\n2,3')
+        kwargs = dict(delimiter=",", missing_values="N/A", names=True)
+        self.assert_deprecated(lambda: recfromcsv(data_gen()))
+        self.assert_deprecated(lambda: recfromtxt(data_gen(), **kwargs))
+
+        self.assert_deprecated(lambda: disp("test"))
+        self.assert_deprecated(lambda: get_array_wrap())
+        self.assert_deprecated(lambda: maximum_sctype(int))
