+++ numpy/lib/function_base.py
@@ -1,10 +1,10 @@
+import builtins
 import collections.abc
 import functools
 import re
 import sys
 import warnings
 
-from .._utils import set_module
 import numpy as np
 import numpy.core.numeric as _nx
 from numpy.core import transpose
@@ -28,8 +28,7 @@ from numpy.core.multiarray import (
     interp as compiled_interp, interp_complex as compiled_interp_complex
     )
 from numpy.core.umath import _add_newdoc_ufunc as add_newdoc_ufunc
-
-import builtins
+from numpy._utils import deprecate, set_module
 
 # needed in this module for compatibility
 from numpy.lib.histograms import histogram, histogramdd  # noqa: F401
@@ -1951,6 +1950,7 @@ def place(arr, mask, vals):
     return _place(arr, mask, vals)
 
 
+@deprecate
 def disp(mesg, device=None, linefeed=True):
     """
     Display a message on a device.
