+++ numpy/ma/tests/test_core.py
@@ -26,7 +26,7 @@ from numpy.testing import (
     )
 from numpy.testing._private.utils import requires_memory
 from numpy import ndarray
-from numpy._utils._convertions import asbytes
+from numpy._utils import asbytes
 from numpy.ma.testutils import (
     assert_, assert_array_equal, assert_equal, assert_almost_equal,
     assert_equal_records, fail_if_equal, assert_not_equal,
