+++ numpy/typing/tests/data/fail/npyio.pyi
@@ -24,7 +24,3 @@ np.savez_compressed(str_file, AR_i8)  # E: incompatible type
 np.loadtxt(bytes_path)  # E: incompatible type
 
 np.fromregex(bytes_path, ".", np.int64)  # E: No overload variant
-
-np.recfromtxt(bytes_path)  # E: incompatible type
-
-np.recfromcsv(bytes_path)  # E: incompatible type
