+++ numpy/typing/tests/data/reveal/numerictypes.pyi
@@ -1,8 +1,5 @@
 import numpy as np
 
-reveal_type(np.maximum_sctype(np.float64))  # E: Type[{float64}]
-reveal_type(np.maximum_sctype("f8"))  # E: Type[Any]
-
 reveal_type(np.issctype(np.float64))  # E: bool
 reveal_type(np.issctype("foo"))  # E: Literal[False]
